refactoring - split into separate services
